시연 후 수정본

+++ main/app_HX711.h
@@ -2,7 +2,7 @@
 #define __APP_HX711_H__
 
 #include "esp_log.h"
-void HX711_cal_init(uint8_t cal);
+void HX711_cal_init(uint16_t cal);
 bool HX711_init(void);
 void HX711_Sensing(void);
 float loadcell_data_get(void);
